app/requirements.txt
torch
torchvision
fastapi[standard]
pillow